+++ node/CHANGELOG.md
@@ -50,7 +50,7 @@ All notable changes to this project will be documented in this file.  The format
 * Shut down SSE event streams gracefully.
 * Limit the maximum number of clients connected to the event stream server via the `[event_stream_server][max_concurrent_subscribers]` config option.
 * Avoid emitting duplicate events in the event stream.
-* Incorrect representation of BlockIdentifier in the Open-RPC schema.
+* Change `BlockIdentifier` params in the Open-RPC schema to be optional.